+++ mediation/LiftoffMonetize/source/android-library/liftoffmonetize/src/main/java/com/google/unity/mediation/liftoffmonetize/VungleUnityInterstitialExtrasBuilder.java
@@ -14,6 +14,8 @@
 
 package com.google.unity.mediation.liftoffmonetize;
 
+import androidx.annotation.NonNull;
+import com.google.android.gms.ads.mediation.MediationExtrasReceiver;
 import com.vungle.mediation.VungleInterstitialAdapter;
 
 /**
@@ -21,8 +23,9 @@ import com.vungle.mediation.VungleInterstitialAdapter;
  */
 public class VungleUnityInterstitialExtrasBuilder extends VungleUnityExtrasBuilder {
 
+  @NonNull
   @Override
-  public Class getAdapterClass() {
+  public Class<? extends MediationExtrasReceiver> getAdapterClass() {
     return VungleInterstitialAdapter.class;
   }
 }